# docs/static/openapi.yaml
openapi: 3.0.3
info:
  title: LCM REST API
  version: 1.29.0
  description: 'Die vollständige REST-API des LCM-Backends (Go/Fiber). Das Web-Frontend nutzt exakt diese
    Endpunkte - es gibt keine separate interne API.


    **Authentifizierung** (zwei gleichwertige Wege, siehe Sicherheitsschemas):

    - `Authorization: Bearer <jwt>` - Browser-/User-Sessions, ausgestellt von `POST /auth/login` (bzw.
    `/auth/login/2fa` bei aktivierter 2FA).

    - `X-API-Key: <key>` - Service-zu-Service-Zugriff mit einem unter *Einstellungen → API-Keys* erzeugten,
    auf Scopes eingeschränkten Schlüssel.


    Berechtigungen werden pro Endpunkt per RBAC-Scope geprüft (in der Beschreibung als `bereich:read`/`bereich:write`
    vermerkt); die sichtbaren Daten werden bei Servern zusätzlich auf die Gruppen des jeweiligen Benutzers
    eingeschränkt.


    Diese Referenz ist automatisch aus den Routen- und Handler-Definitionen des Quellcodes erzeugt (`internal/api/router/router.go`
    + Controller) - Pfade, Methoden und Berechtigungen sind exakt der aktuelle Stand. Nicht jeder Endpunkt
    hat ein vollständig dokumentiertes Response-Schema; die tatsächliche Antwort richtet sich nach dem
    jeweiligen Domain-Modell (siehe Architektur).'
servers:
- url: /api/v1
  description: Relativ zur LCM-Instanz, z. B. https://lcm.example.com/api/v1
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session-Token aus POST /auth/login.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Service-API-Key aus Einstellungen → API-Keys.
...
paths:
...
  /api/v1/privilege-profiles:
    get:
      tags:
      - Berechtigungsprofile
      summary: Alle Berechtigungsprofile.
      operationId: profileCtrl_List
      description: 'Liefert die Profile samt ihrer sudo-, Editier- und Verzeichnisregeln sowie der verwendeten Regelbausteine.

        

        

        Benötigte Berechtigung: `profiles:read`.'
      security:
      - sessionAuth: []
      - apiKeyAuth: []
      responses:
        '200':
          description: Erfolgreich.
        '401':
          description: Nicht angemeldet.
        '403':
          description: Keine Berechtigung.
    post:
      tags:
      - Berechtigungsprofile
      summary: Berechtigungsprofil anlegen.
      operationId: profileCtrl_Create
      description: 'Legt ein Profil an. Regeln, die faktisch Root-Rechte vergeben (Shell-Metazeichen, Wildcards, root-aequivalente Binaries), werden mit 422 abgelehnt - es sei denn, die Regel setzt allow_root_equivalent ausdruecklich.

        

        

        Benötigte Berechtigung: `profiles:write`.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                slug:
                  type: string
                description:
                  type: string
                account_type:
                  type: string
                sudo_rules:
                  type: array
                  items:
                    type: object
                edit_rules:
                  type: array
                  items:
                    type: object
                path_rules:
                  type: array
                  items:
                    type: object
                block_uses:
                  type: array
                  items:
                    type: object
      security:
      - sessionAuth: []
      - apiKeyAuth: []
      responses:
        '200':
          description: Erfolgreich.
        '401':
          description: Nicht angemeldet.
        '403':
          description: Keine Berechtigung.
  /api/v1/privilege-profiles/{id}:
    delete:
      tags:
      - Berechtigungsprofile
      summary: Berechtigungsprofil loeschen.
      operationId: profileCtrl_Delete
      description: 'Loescht ein Profil. Ein Profil, das noch zugewiesen ist, wird mit 409 abgelehnt; mitgelieferte Profile sind geschuetzt (403).

        

        

        Benötigte Berechtigung: `profiles:write`.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - sessionAuth: []
      - apiKeyAuth: []
      responses:
        '200':
          description: Erfolgreich.
        '401':
          description: Nicht angemeldet.
        '403':
          description: Keine Berechtigung.
    get:
      tags:
      - Berechtigungsprofile
      summary: Ein Berechtigungsprofil.
      operationId: profileCtrl_Get
      description: 'Liefert ein einzelnes Profil samt aller Regeln.

        

        

        Benötigte Berechtigung: `profiles:read`.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      security:
      - sessionAuth: []
      - apiKeyAuth: []
      responses:
        '200':
          description: Erfolgreich.
        '401':
          description: Nicht angemeldet.
        '403':
          description: Keine Berechtigung.
    patch:
      tags:
      - Berechtigungsprofile
      summary: Berechtigungsprofil aendern.
      operationId: profileCtrl_Update
      description: 'Aendert Kopfdaten und Regeln. Die Regeln werden ersetzt, nicht zusammengefuehrt: Ein Profil beschreibt einen Soll-Zustand.

        

        

        Benötigte Berechtigung: `profiles:write`.'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                account_type:
                  type: string
                sudo_rules:
                  type: array
                  items:
                    type: object
                edit_rules:
                  type: array
                  items:
                    type: object
                path_rules:
                  type: array
                  items:
                    type: object
                block_uses:
                  type: array
                  items:
                    type: object
      security:
      - sessionAuth: []
      - apiKeyAuth: []
      responses:
        '200':
          description: Erfolgreich.
        '401':
          description: Nicht angemeldet.
        '403':
          description: Keine Berechtigung.
...
  /api/v1/users/{id}/profile:
    patch:
      tags:
      - Benutzer (LCM-Login)
      summary: Admins (users:write) dürfen jeden bearbeiten; jeder eingeloggte User darf sein EIGENES
        Profil bearbeiten.
      operationId: userCtrl_UpdateProfile
      description: Admins (users:write) dürfen jeden bearbeiten; jeder eingeloggte User darf sein EIGENES
        Profil bearbeiten.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
      security:
      - sessionAuth: []
      - apiKeyAuth: []
      responses:
        '200':
          description: Erfolgreich.
        '401':
          description: Nicht angemeldet.
        '403':
          description: Keine Berechtigung.
...